+++ RELEASE-NOTES.md
@@ -91,6 +91,7 @@ Initial version with:
 * Account Managing > Lock a sub-account
 * Account Managing > Unlock a sub-account
 * Account Managing > Transfer credits
+* Account Managing > Data retention > Get
 
 ### Upgrade Steps
 * [NO ACTION REQUIRED]
@@ -108,6 +109,7 @@ Initial version with:
 * Account Managing > Lock a sub-account
 * Account Managing > Unlock a sub-account
 * Account Managing > Transfer credits
+* Account Managing > Data retention > Get
 
 ### Bug Fixes
 * Fixed recipients bugs on campaigns, create tokens, create webhook, HLR lookup, update webhook